move custom deletion tool for cleanup jobs to utils
