+++ awx/main/utils/deletion.py
